generalize `main` over I/O types
