+++ a3/Cargo.toml
@@ -10,6 +10,4 @@ description = "Solution for SPCC Kickstart Problem A3"
 
 [dependencies]
 unicode-segmentation = "1.9"
-
-[dev-dependencies]
-test_run_bin = { path = "../test_run_bin" }
+kickstart_macros = { path = "../kickstart_macros" }
