+++ a5/Cargo.toml
@@ -10,8 +10,7 @@ description = "Solution for SPCC Kickstart Problem A5"
 
 [dependencies]
 read_u32 = { path = "../read_u32" }
+kickstart_macros = { path = "../kickstart_macros" }
 
 [dev-dependencies]
-test_run_bin = { path = "../test_run_bin" }
-predicates = "2.1"
 quickcheck = "1"
