+++ a5/tests/basic.rs
@@ -1,6 +1,4 @@
-#![cfg(not(miri))] // miri can't run other executables
-
-test_run_bin::initialize!();
+kickstart_macros::fn_run!(a5);
 
 #[test]
 fn sample() {
