+++ b5/src/lib.rs
@@ -1,53 +1,63 @@
-/// Validates the 9x9 sudoku board, with input numbers from 0 to 8.
-#[must_use]
-pub fn validate(board: &[[u8; 9]; 9]) -> bool {
-    // The values in all groups must be a permutation of 0, 1, 2, 3, 4, 5, 6, 7 and 8.
-    #[rustfmt::skip]
-    const GROUPS: [[[usize; 2]; 9]; 27] = [
-        // Rows
-        [[0, 0], [0, 1], [0, 2], [0, 3], [0, 4], [0, 5], [0, 6], [0, 7], [0, 8]],
-        [[1, 0], [1, 1], [1, 2], [1, 3], [1, 4], [1, 5], [1, 6], [1, 7], [1, 8]],
-        [[2, 0], [2, 1], [2, 2], [2, 3], [2, 4], [2, 5], [2, 6], [2, 7], [2, 8]],
-        [[3, 0], [3, 1], [3, 2], [3, 3], [3, 4], [3, 5], [3, 6], [3, 7], [3, 8]],
-        [[4, 0], [4, 1], [4, 2], [4, 3], [4, 4], [4, 5], [4, 6], [4, 7], [4, 8]],
-        [[5, 0], [5, 1], [5, 2], [5, 3], [5, 4], [5, 5], [5, 6], [5, 7], [5, 8]],
-        [[6, 0], [6, 1], [6, 2], [6, 3], [6, 4], [6, 5], [6, 6], [6, 7], [6, 8]],
-        [[7, 0], [7, 1], [7, 2], [7, 3], [7, 4], [7, 5], [7, 6], [7, 7], [7, 8]],
-        [[8, 0], [8, 1], [8, 2], [8, 3], [8, 4], [8, 5], [8, 6], [8, 7], [8, 8]],
+use std::io::prelude::*;
 
-        // Columns
-        [[0, 0], [1, 0], [2, 0], [3, 0], [4, 0], [5, 0], [6, 0], [7, 0], [8, 0]],
-        [[0, 1], [1, 1], [2, 1], [3, 1], [4, 1], [5, 1], [6, 1], [7, 1], [8, 1]],
-        [[0, 2], [1, 2], [2, 2], [3, 2], [4, 2], [5, 2], [6, 2], [7, 2], [8, 2]],
-        [[0, 3], [1, 3], [2, 3], [3, 3], [4, 3], [5, 3], [6, 3], [7, 3], [8, 3]],
-        [[0, 4], [1, 4], [2, 4], [3, 4], [4, 4], [5, 4], [6, 4], [7, 4], [8, 4]],
-        [[0, 5], [1, 5], [2, 5], [3, 5], [4, 5], [5, 5], [6, 5], [7, 5], [8, 5]],
-        [[0, 6], [1, 6], [2, 6], [3, 6], [4, 6], [5, 6], [6, 6], [7, 6], [8, 6]],
-        [[0, 7], [1, 7], [2, 7], [3, 7], [4, 7], [5, 7], [6, 7], [7, 7], [8, 7]],
-        [[0, 8], [1, 8], [2, 8], [3, 8], [4, 8], [5, 8], [6, 8], [7, 8], [8, 8]],
+mod validate;
 
-        // Grids
-        [[0, 0], [0, 1], [0, 2], [1, 0], [1, 1], [1, 2], [2, 0], [2, 1], [2, 2]],
-        [[3, 0], [3, 1], [3, 2], [4, 0], [4, 1], [4, 2], [5, 0], [5, 1], [5, 2]],
-        [[6, 0], [6, 1], [6, 2], [7, 0], [7, 1], [7, 2], [8, 0], [8, 1], [8, 2]],
-        [[0, 3], [0, 4], [0, 5], [1, 3], [1, 4], [1, 5], [2, 3], [2, 4], [2, 5]],
-        [[3, 3], [3, 4], [3, 5], [4, 3], [4, 4], [4, 5], [5, 3], [5, 4], [5, 5]],
-        [[6, 3], [6, 4], [6, 5], [7, 3], [7, 4], [7, 5], [8, 3], [8, 4], [8, 5]],
-        [[0, 6], [0, 7], [0, 8], [1, 6], [1, 7], [1, 8], [2, 6], [2, 7], [2, 8]],
-        [[3, 6], [3, 7], [3, 8], [4, 6], [4, 7], [4, 8], [5, 6], [5, 7], [5, 8]],
-        [[6, 6], [6, 7], [6, 8], [7, 6], [7, 7], [7, 8], [8, 6], [8, 7], [8, 8]],
-    ];
+pub fn main_impl(input: impl BufRead, mut output: impl Write) {
+    match input_sudoku_board(input) {
+        Some(board) if validate::validate(&board) => {
+            writeln!(output, "1")
+        }
+        _invalid => {
+            writeln!(output, "0")
+        }
+    }
+    .expect("An output error occured");
+}
 
-    GROUPS.into_iter().all(|group| {
-        group
-            .into_iter()
-            // `value_exists_array[i]` is true when `i` exists in the group.
-            // When the entire `value_exists_array` is true, then all numbers from 0 to 8 exist.
-            .fold([false; 9], |mut value_exists_table, cell_index| {
-                let value = board[cell_index[0]][cell_index[1]];
-                value_exists_table[usize::from(value)] = true;
-                value_exists_table
-            })
-            == [true; 9]
-    })
+// Input routine.
+// Returns a two-dimensional array containing numbers from 0 to 8.
+//
+// This part completes all validation required for Subtask 1.
+fn input_sudoku_board(mut input: impl BufRead) -> Option<[[u8; 9]; 9]> {
+    // An input buffer.
+    let mut buf = Vec::new();
+
+    // Read the first line into the buffer.
+    input
+        .read_until(b'\n', &mut buf)
+        .expect("An input error occured");
+
+    // The first line, which contains the width and height, must be "9 9" to be valid.
+    match buf.as_slice() {
+        b"9 9\n" | b"9 9\r\n" => {}
+        _ => return None,
+    }
+
+    let mut result_array = [[0u8; 9]; 9];
+
+    // Read the following 9 lines of 9 characters into the buffer.
+    for row in &mut result_array {
+        // Read one row into the buffer; it is guaranteed to have 9 characters.
+        buf.clear();
+        input
+            .read_until(b'\n', &mut buf)
+            .expect("An input error occured");
+
+        // Check that the first 9 bytes are valid ASCII digits from '1' to '9'.
+        for (cell, &byte) in std::iter::zip(row, &buf) {
+            match byte {
+                // Subtracting the digit by b'1' gives a number from 0 to 8 (corresponding to '1' to '9').
+                digit @ b'1'..=b'9' => *cell = digit - b'1',
+                _other => return None,
+            }
+        }
+
+        // If there are any additional bytes, this is an incorrect row.
+        match &buf[9..] {
+            b"" | b"\n" | b"\r\n" => {}
+            _additional => return None,
+        }
+    }
+
+    Some(result_array)
 }
