+++ kickstart_macros/Cargo.toml
@@ -1,9 +1,6 @@
 [package]
-name = "test_run_bin"
+name = "kickstart_macros"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-assert_cmd = "2.0"
